Move SQLHelper back from db module
